+++ src/appVersion.ts
@@ -13,7 +13,7 @@ import inBrowser from './inBrowser'
  * ```
  * @since 5.1.0
  * @param appName - app name
- * @param ua - ua or any ua like string, may not be passed, default is navigator.userAgent
+ * @param ua - ua or any ua like string, allowed to be undefined, default is navigator.userAgent
  * @param ignoreCase - whether to ignore case
  * @return string|null
  */